% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portuguese.R
\name{as.portuguese}
\alias{as.portuguese}
\alias{portuguese.default}
\alias{portuguese.numeric}
\alias{as.numeric.portuguese}
\alias{print.portuguese}
\alias{sort.portuguese}
\alias{rep.portuguese}
\alias{[.portuguese}
\alias{format.portuguese}
\title{Generic functions and methods for S3 class english}
\usage{
as.portuguese(x, ...)

\method{portuguese}{default}(x, ...)

\method{portuguese}{numeric}(x, BR, ...)

\method{as.numeric}{portuguese}(x, ...)

\method{print}{portuguese}(x, ...)

\method{sort}{portuguese}(x, decreasing = FALSE, ...)

\method{rep}{portuguese}(x, ...)

\method{[}{portuguese}(x, i)

\method{format}{portuguese}(x, ...)
}
\arguments{
\item{x}{Any numeric object.  If the components are not integers, they are
rounded.  In the case of \code{portuguese.default} this may be an
object of any class, but will always result in an error.}

\item{...}{For \code{rep}, and \code{format} arguments passed on to other
methods.  Ignored in all other cases.}

\item{BR}{A logical flag, should Brazil-style numbers be given (\code{TRUE})
or Portugal style (\code{FALSE})? If missing, a default value is
chosen as \code{TRUE} unless the locale in use is guessed to be an
Portuguese locale of Portugal.}

\item{decreasing}{logical: should the object be sorted in decreasing order?}

\item{i}{Any allowable form of index vector.}
}
\value{
An object of class \code{portuguese}, if possible, or an error message if
    not.

A numeric vector of class \code{portuguese}.

The original object, invisibly.

An object of class \code{portuguese} with its components in numerically
    sorted order.
}
\description{
The functions \code{as.portuguese} and \code{portuguese} are fully
   equivalent generic constructor functions for the S3 class
   \code{portuguese}, which allows numeric objects to be represented in
   the form of their customary Portuguese expression.

Provides a \code{print} method for objects of class \code{portuguese}.
}
\details{
The function \code{portuguese} is a generic constructor function for
    objects of class \code{portuguese}.  The methods either mark the object
    as having the appropriate class, or, in the case of
    \code{portuguese.default}, result in an error message.  The function
    \code{as.portuguese} is provided as a corresponding function to
    \code{as.roman} in the \code{utils} package.  The method
    \code{portuguese.portuguese} simply allows the constructor function to act
    as a coercion and to have no effect on objects already of the class.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Generic functions and methods for S3 class english

\item \code{numeric}: Convert an portuguese object back to class numeric

\item \code{portuguese}: Convert an Portuguese object back to class numeric

This a convenience function that simply removes the S3 class attribute
   from and object of class \code{portuguese}.

\item \code{portuguese}: A print method for objects of class portuguese

\item \code{portuguese}: A sorting method for objects of class  portuguese

A method for the generic function \code{sort} which allows numeric
    objects of class \code{portuguese} to be sorted.

\item \code{portuguese}: class portuguese

\item \code{portuguese}: Old method

\item \code{portuguese}: Format for portuguese
}}

