% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/english.R
\name{Ops.english}
\alias{Ops.english}
\title{English Arithmetic}
\usage{
\method{Ops}{english}(e1, e2)
}
\arguments{
\item{e1}{Numeric vector of object of class \code{"english"}.}

\item{e2}{Numeric vector of object of class \code{"english"}.}
}
\value{
Numeric vector of class \code{"english"}
}
\description{
Allows arithmetic operations on \code{"english"} class objects.  To make sense
the operation should return an integer value.
}
\examples{
english(1:10)^2 + 1:10
english(100) + (-5):5
}
