% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portuguese.R
\name{Ops.portuguese}
\alias{Ops.portuguese}
\title{Arithmetic operations with objects of class portuguse}
\usage{
\method{Ops}{portuguese}(e1, e2)
}
\arguments{
\item{e1}{Numeric vectors, of which at least one must be of class \code{portuguese}.}

\item{e2}{Numeric vectors, of which at least one must be of class \code{portuguese}.}
}
\value{
A numeric vector of class \code{portuguese}.
}
\description{
Allows arithmetic with objects of class \code{portuguese}.
}
