% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words.R
\name{words}
\alias{words}
\alias{Words}
\title{Express Numbers in Words}
\usage{
words(x)

Words(x)
}
\arguments{
\item{x}{A numeric vector, usually integer}
}
\value{
A character string vector with the numbers expressed
        in English words.
}
\description{
Convert numerical objects to Enghish character strings.
A convenience function for use mainly with \code{RMarkdown}
in-text inserts.  The capitalized version, \code{Words},
makes the initial letter of the result upper case.
}
\examples{
cat("The Duke of York had ", words(10006), " men.\\n", sep = "")
cat("How many did he have? ", Words(10006), ".\\n", sep = "")
}
