% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal.R
\name{ordinal}
\alias{ordinal}
\alias{ordinal.english}
\alias{ordinal.numeric}
\alias{ordinal.character}
\alias{print.ordinal}
\title{Ordinal Numbers}
\usage{
ordinal(x, ...)

\method{ordinal}{english}(x, ...)

\method{ordinal}{numeric}(x, ...)

\method{ordinal}{character}(x, ...)

\method{print}{ordinal}(x, ...)
}
\arguments{
\item{x}{A numeric vector, usually integer, or an object of class "ordinal"}

\item{...}{Ignored. Included only for compatibility.}
}
\value{
A character string vector of ordinal versions of the number, with S3 class "ordinal"
}
\description{
Generates character strings of the ordinal version of numbers in English words.
}
\examples{
ordinal(1:12)
}
